Added support for file source

+++ include/data.h
@@ -7,15 +7,15 @@
 
 #include <stdbool.h>
 
-void join_triple_copy(const data_t *src1, uint8_t index1,
-                    const data_t *src2, uint8_t index2, data_t *dest);
+void join_triple_copy(const data_t *src1, uint32_t index1,
+                    const data_t *src2, uint32_t index2, data_t *dest);
 
-bool join_check(const data_t *src1, uint8_t index1,
-                    const data_t *src2, uint8_t index2, join_params_t check);
+bool join_check(const data_t *src1, uint32_t index1,
+                    const data_t *src2, uint32_t index2, join_params_t check);
 
-void triple_copy(const data_t *src, uint8_t index, data_t *dest);
+void triple_copy(const data_t *src, uint32_t index, data_t *dest);
 
-bool filter_check(const data_t *src, uint8_t index, filter_params_t check);
+bool filter_check(const data_t *src, uint32_t index, filter_params_t check);
 
 void free_data(data_t *data);
